Fill in basic structure of readme
And fill in stage-0 as baseline for which to add Bazel.
Note that this has the rust sources in a `backend` directory.
Hopefully, this better represents scenarios in which someone
would reach for bazel and works around issues of vendoring
and running bazel from the root directory
(see https://github.com/bazelbuild/rules_rust/issues/1341)

* Add stage 0

    From https://github.com/tokio-rs/axum/tree/82e8eaeee4503850f1ce36cedd46b4afc9ae4850/examples/readme

* Fill in rough structure of readme